First
From the concept of Bidirectional visitor  counter,it is easy to minimize power consumption  and also it is low in cost.Electricity usage is minimized by the automatic switching ON/OFF of lights and Fans based on visitor’s count. If any fire breakout,the system generates alarm and pumps up water automatically to reduce huge loss.